fixed 15-point question for 12 foils
AMC seems to limit the width of the options.  So the middle values increment by 2.  I have found that when I grade I may want to give half credit, so I made the middle 7.5.  But I took out the values nearest it.  I find that at the high end, I often want to take off a point or two.  At the low end I may want to give credit for the small bits someone did get correct.  But in the middle, broader steps are possible.